requirements.txt
Django>=6.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
razorpay==1.4.2
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
waitress==3.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-colorfield==0.11.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
